Improve webpack configuration (#9225)
Improve the webpack configuration to make building LORIS more memory and time efficient. The changes of this PR are described in https://github.com/aces/Loris/issues/9218#issuecomment-2079578910

Resolves #9218
